OpenGLRenderer: fix scissor rect calculation in __render()
When __context3D.__backBufferWantsBestResolution is true, need to account for __pixelRatio, or the scissor rect will be too large or too small, depending on the __pixelRatio.

Followup to commit da0463a39863cf9caa86fe2d9716d7579251da5a, which includes some test code in a comment that includes scrollRect, filters, and Starling masks, which all affect the scissor rect.